src/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
chromadb==0.4.15
sentence-transformers==2.2.2
google-generativeai==0.3.2
pymupdf==1.23.8
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.4
huggingface-hub==0.17.3
transformers==4.35.0